tools/iot_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
httpx==0.27.2
